gai-lib/requirements.txt
pillow==10.2
#unstructured[all-docs]==0.12.0
markdownify
httpx
openai
anthropic
python-dotenv
langchain
colorlog==6.8.2
